+++ src/comm/publish/tests.rs
@@ -2951,10 +2951,10 @@ mod fixtures {
             error_code: 1,
             client_realm: "client_realm".to_string(),
             cname_type: 1,
-            client_name: vec!["client_name".to_string()],
+            cname: vec!["client_name".to_string()],
             realm: "realm".to_string(),
             sname_type: 1,
-            service_name: vec!["service_name".to_string()],
+            sname: vec!["service_name".to_string()],
         };
 
         bincode::serialize(&kerberos_body).unwrap()